+++ numpy/lib/__init__.py
@@ -1,11 +1,14 @@
 from __future__ import division, absolute_import, print_function
 
+import math
+
 from .info import __doc__
 from numpy.version import version as __version__
 
 from .type_check import *
 from .index_tricks import *
 from .function_base import *
+from .nanfunctions import *
 from .shape_base import *
 from .stride_tricks import *
 from .twodim_base import *
@@ -18,7 +21,6 @@ from .utils import *
 from .arraysetops import *
 from .npyio import *
 from .financial import *
-import math
 from .arrayterator import *
 from .arraypad import *
 
@@ -36,6 +38,7 @@ __all__ += utils.__all__
 __all__ += arraysetops.__all__
 __all__ += npyio.__all__
 __all__ += financial.__all__
+__all__ += nanfunctions.__all__
 
 from numpy.testing import Tester
 test = Tester().test
